+++ packages/cli/src/commands/index.ts
@@ -140,6 +140,32 @@ export function registerScriptCommand(program: Command) {
     .action(lazy(() => import('./export-dynamic-plugin').then(m => m.command)));
 
   command
+    .command('package-dynamic-plugins')
+    .description(
+      'Package up exported dynamic plugins as container image for deployment',
+    )
+    .option(
+      '--force-export',
+      'Regenerate the dist-dynamic folder for each plugin even if it already exists',
+    )
+    .option(
+      '--preserve-temp-dir',
+      'Leave the temporary staging directory on the filesystem instead of deleting it',
+    )
+    .requiredOption(
+      '-t, --tag <tag>',
+      'Tag name to use when building the plugin registry image',
+    )
+    .option(
+      '--use-docker',
+      'By defult, the command uses podman to build the container image. Use this flag to use docker instead.',
+      false,
+    )
+    .action(
+      lazy(() => import('./package-dynamic-plugins').then(m => m.command)),
+    );
+
+  command
     .command('schema')
     .description('Print configuration schema for a package')
     .action(lazy(() => import('./schema').then(m => m.default)));