+++ plugins/shared-react/src/components/index.ts
@@ -12,3 +12,4 @@ export {
   DrawerContext,
   useDrawer,
 } from './common/DrawerContext';
+export { Status } from './common/Status';