+++ plugins/tekton/src/components/PipelineRunList/PipelineRunList.test.tsx
@@ -29,6 +29,15 @@ const mockUsePermission = usePermission as jest.MockedFunction<
   typeof usePermission
 >;
 
+jest.mock('@material-ui/styles', () => ({
+  ...jest.requireActual('@material-ui/styles'),
+  makeStyles: () => (_theme: any) => {
+    return {
+      ok: 'ok',
+    };
+  },
+}));
+
 describe('PipelineRunList', () => {
   beforeEach(() => {
     mockUsePermission.mockReturnValue({ loading: false, allowed: true });